requirements/api.txt
# ============================================================================
# Python Dependencies for Inference API Service
# ============================================================================

# Web Framework
fastapi>=0.105.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Core ML Libraries (for inference)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# MLflow Integration (for loading models)
mlflow>=2.9.0
boto3>=1.29.0  # For S3/MinIO artifact storage

# Data Connectors
kafka-python>=2.0.2
cassandra-driver>=3.28.0
redis>=5.0.0

# Serialization
pyarrow>=14.0.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# HTTP Client
httpx>=0.25.0

# Caching
cachetools>=5.3.0

# Rate Limiting
slowapi>=0.1.9

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

